+++ UI/vue3-antd/components/DataActionLogsModal.js
@@ -7,7 +7,6 @@ import { Button, Modal, Table, Pagination, Select, SelectOption } from 'ant-desi
  * @param params.application_id 应用id
  * @param params.function_type 模块类型
  * @param params.token 令牌
- * @param params.title 标题
  * @param getDataActionLog 获取操作日志的request方法
  * @param getLogSearchParams 获取日志查询的类型
  * @param langTexts 国际化文字
@@ -19,6 +18,10 @@ export default function (params, getDataActionLog, getLogSearchParams, langTexts
     props: {
       data_id: {
         type: [Number, String],
+        required: true,
+      },
+      title: {
+        type: String,
       },
     },
     setup(props, { slots }) {
@@ -123,7 +126,7 @@ export default function (params, getDataActionLog, getLogSearchParams, langTexts
                 },
               },
               {
-                title: () => (params.title ? params.title : langTexts.actionLog),
+                title: () => (props.title ? props.title : langTexts.actionLog),
                 footer: () =>
                   h(
                     Button,