+++ example/vue3-antdv/src/App.vue
@@ -1,30 +1,51 @@
 <template>
-  <div class="box">
-    <div class="nav">
-      <a-menu :selectedKeys="['home']" mode="horizontal">
-        <a-menu-item key="home">
-          <router-link to="/home">
-            api功能操作介绍
-          </router-link>
-        </a-menu-item>
-        <a-menu-item @click="logoutHandler">
-          退出登录
-        </a-menu-item>
-      </a-menu>
+  <a-config-provider :locale="zhCN">
+    <div class="box">
+      <a-spin :spinning="spinning">
+        <div class="nav">
+          <div>
+            用户名：{{userInfo?.name}}
+          </div>
+          <a-button @click="logoutHandler">退出登录</a-button>
+        </div>
+        <router-view/>
+      </a-spin>
     </div>
-    <router-view/>
-  </div>
+  </a-config-provider>
 </template>
 
 <script setup>
 import { IAM } from '@/utils/iamUtils'
+import { useStore } from 'vuex'
+import { computed, ref } from 'vue'
+
+import zhCN from 'ant-design-vue/es/locale/zh_CN'
+
+const store = useStore()
+const spinning = ref(false)
+
+const userInfo = computed(() => {
+  return store.state.userInfo
+})
 
 const logoutHandler = () => {
   const b = window.confirm('您确定退出登录吗？')
   if (b) {
-    IAM.signOut()
+    spinning.value = true
+    IAM.signOut({
+      logoutBeforeCallback () {
+        alert('马上就要退出了，别激动！')
+      },
+      logoutSuccessCallback () {
+        alert('退出登录成功！')
+      },
+      logoutErrorCallback () {
+        alert('退出登录错误的回调')
+      }
+    })
   }
 }
+
 </script>
 
 <style lang="scss">
@@ -33,6 +54,11 @@ const logoutHandler = () => {
   width: 1000px;
   margin: 0 auto;
   overflow: auto;
+}
 
+.nav{
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
 }
 </style>