+++ example/vue3-antdv/src/views/Home.vue
@@ -1,12 +1,38 @@
 <template>
-  <div class="home">
-    {{ environment }}
+  <div class="home" style="margin-top: 20px;">
+    <a-collapse v-model:activeKey="activeKey">
+      <a-collapse-panel key="1" header="获取用户信息（在iam系统中的，账号信息和所拥有系统的权限信息）">
+        <GetUserInfo/>
+      </a-collapse-panel>
+      <a-collapse-panel key="2" header="获取认证信息（auth2.0返回的token等信息）" >
+        <GetAuthInfo/>
+      </a-collapse-panel>
+      <a-collapse-panel key="3" header="获取权限信息（当前系统-iam管理台的系统的权限在iam中的菜单按钮权限）" >
+        <GetPermissionInfo/>
+      </a-collapse-panel>
+      <a-collapse-panel key="4" header="获取首页地址（当前环境下的对应的iam-dashboard的地址，注意：iam 没有开发环境，开发环境、测试环境均对应iam的测试环境）" >
+        <GetHomePath/>
+      </a-collapse-panel>
+      <a-collapse-panel key="5" header="检测错误的响应json，提供检测是否存在token失效，设备被挤，被强制退出" >
+        <CheckErrorString/>
+      </a-collapse-panel>
+      <a-collapse-panel key="6" header="日志查询组件功能" >
+        <ActionLog/>
+      </a-collapse-panel>
+    </a-collapse>
+
   </div>
 </template>
 
 <script setup>
-import { IAM } from '../utils/iamUtils'
+import { ref } from 'vue'
 
-// 当前环境
-const environment = IAM
+import GetUserInfo from '../components/getUserInfo/Index.vue'
+import GetAuthInfo from '../components/getAuthInfo/Index.vue'
+import GetPermissionInfo from '../components/getPermissionInfo/Index.vue'
+import GetHomePath from '../components/getHomePath/Index.vue'
+import CheckErrorString from '../components/checkErrorString/Index.vue'
+import ActionLog from '../components/actionLogComponent/Index.vue'
+
+const activeKey = ref('1')
 </script>